refactor(iocuak-core): move getBinding to binding module
