+++ packages/iocuak-core/src/task/actions/createInstancesByTag.spec.ts
@@ -1,7 +1,5 @@
 import { afterAll, beforeAll, describe, expect, it, jest } from '@jest/globals';
 
-jest.mock('./getBinding');
-
 import { Tag } from '@cuaklabs/iocuak-common';
 import { Binding } from '@cuaklabs/iocuak-models';
